+++ python/cudf/tests/test_dataframe.py
@@ -494,8 +494,8 @@ def test_dataframe_iloc(nelem):
 
 
 @pytest.mark.xfail(
-    raises=NotImplementedError,
-    reason="cudf columnar iloc not supported"
+    raises=AssertionError,
+    reason="Series.index are different"
 )
 def test_dataframe_iloc_tuple():
     gdf = DataFrame()
@@ -508,11 +508,11 @@ def test_dataframe_iloc_tuple():
     pdf['a'] = ha
     pdf['b'] = hb
 
-    def assert_col(g, p):
-        np.testing.assert_equal(g['a'].to_array(), p['a'])
-        np.testing.assert_equal(g['b'].to_array(), p['b'])
+    # We don't support passing the column names into the index quite yet
+    got = gdf.iloc[1, [1]]
+    expect = pdf.iloc[1, [1]]
 
-    assert_col(gdf.iloc[1, 2], pdf.iloc[1, 2])
+    assert_eq(got, expect)
 
 
 @pytest.mark.xfail(
@@ -693,6 +693,25 @@ def test_dataframe_dtypes():
     df = DataFrame([(k, np.ones(10, dtype=v))
                     for k, v in dtypes.iteritems()])
     assert df.dtypes.equals(dtypes)
+
+
+def test_dataframe_add_col_to_object_dataframe():
+    # Test for adding column to an empty object dataframe
+    cols = ['a', 'b', 'c']
+    df = pd.DataFrame(columns=cols, dtype='str')
+
+    data = {k: v for (k, v) in zip(cols, [['a'] for _ in cols])}
+
+    gdf = DataFrame(data)
+    gdf = gdf[:0]
+
+    assert gdf.dtypes.equals(df.dtypes)
+    gdf['a'] = [1]
+    df['a'] = [10]
+    assert gdf.dtypes.equals(df.dtypes)
+    gdf['b'] = [1.0]
+    df['b'] = [10.0]
+    assert gdf.dtypes.equals(df.dtypes)
 
 
 def test_dataframe_dir_and_getattr():
@@ -1202,18 +1221,22 @@ def test_to_from_arrow_nulls(data_type):
         s1 = pa.array([1, None, 3, None, 5], type=data_type)
     gs1 = gd.Series.from_arrow(s1)
     assert isinstance(gs1, gd.Series)
+    # We have 64B padded buffers for nulls whereas Arrow returns a minimal
+    # number of bytes, so only check the first byte in this case
     np.testing.assert_array_equal(
-        np.array(s1.buffers()[0]),
-        gs1.nullmask.to_array()
+        np.array(s1.buffers()[0])[0],
+        gs1.nullmask.to_array()[0]
     )
     assert pa.Array.equals(s1, gs1.to_arrow())
 
     s2 = pa.array([None, None, None, None, None], type=data_type)
     gs2 = gd.Series.from_arrow(s2)
     assert isinstance(gs2, gd.Series)
+    # We have 64B padded buffers for nulls whereas Arrow returns a minimal
+    # number of bytes, so only check the first byte in this case
     np.testing.assert_array_equal(
-        np.array(s2.buffers()[0]),
-        gs2.nullmask.to_array()
+        np.array(s2.buffers()[0])[0],
+        gs2.nullmask.to_array()[0]
     )
     assert pa.Array.equals(s2, gs2.to_arrow())